Ruff fixes and ignores.

+++ qt/src/toga_qt/widgets/timeinput.py
@@ -45,5 +45,5 @@ class TimeInput(Widget):
             value.hour,
             value.minute,
             value.second,
-            int(round(value.microsecond / 1000)),
+            round(value.microsecond / 1000),
         )
